+++ src/main/java/kitchenpos/menugroup/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menugroup.repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
